moved redshift partition tests to test_publish_redshift
